chore(permission): 변수명 변경

+++ src/main/java/com/pooli/permission/domain/entity/PermissionLine.java
@@ -16,7 +16,7 @@ public class PermissionLine {
 
     private Long lineId;
 
-    private Long permissionId;
+    private Integer permissionId;
 
     private Boolean isEnable;
